+++ ergo_std/src/exec.rs
@@ -1,6 +1,12 @@
 //! Execute (possibly with path-lookup) external programs.
 
-use ergo_runtime::abi_stable::{ffi::OsString, std_types::ROption, type_erase::Erased, StableAbi};
+use ergo_runtime::abi_stable::{
+    ffi::OsString,
+    rvec,
+    std_types::{ROption, RString, RVec},
+    type_erase::Erased,
+    StableAbi,
+};
 use ergo_runtime::{
     context::ItemContent,
     depends,
@@ -9,11 +15,11 @@ use ergo_runtime::{
     metadata::Source,
     nsid, traits, try_result,
     type_system::ErgoType,
-    types, Context, Error, Value,
+    types, Context, Value,
 };
-use futures::channel::oneshot::channel;
-use futures::future::{FutureExt, TryFutureExt};
-use futures::stream::StreamExt;
+use futures::future::FutureExt;
+use futures::io::AsyncWriteExt;
+use futures::lock::Mutex;
 use std::process::{Command, Stdio};
 
 /// Strings used for commands and arguments.
@@ -40,6 +46,19 @@ impl ergo_runtime::GetDependenciesConstant for CommandString {
         depends![CommandString::ergo_type(), self]
     }
 }
+
+/// A spawned child process.
+#[derive(Clone, Debug, ErgoType, StableAbi)]
+#[repr(C)]
+pub struct Child {
+    command_string: RString,
+    stdin: Value,
+    stdout: Value,
+    stderr: Value,
+    exit_status: Value,
+}
+
+const CHILD_SUCCESS_INDEX: &'static str = "success";
 
 /// The exit status of a command.
 #[derive(Clone, Debug, Eq, PartialEq, Hash, ErgoType, StableAbi)]
@@ -108,19 +127,23 @@ impl From<ExitStatus> for ergo_runtime::TypedValue<ExitStatus> {
 /// * `Map :env`: A map of Strings to Strings where key-value pairs define the environment
 /// variables to set while executing the program (default empty).
 /// * `Into<Path> :pwd`: The working directory to set while executing the program (default none).
-/// * `Into<ByteStream> :stdin`: The stdin to pipe into the program (default none).
 /// * `:retain-terminal`: If set, the spawned child is kept in the same terminal session.
 ///
-/// This returns a Map with the following indices:
-/// * `:stdout`: The standard output stream from the program, as a `ByteStream`.
-/// * `:stderr`: The standard error stream from the program, as a `ByteStream`.
-/// * `:exit-status`: The exit status of the program, as an `ExitStatus`.
-/// * `:complete`: Waits for the program to complete successfully, evaluating to `Unit` or an
-/// `Error`.
-///
-/// Note that the `complete` value will not evaluate to an error if `exit-status` is used in the
-/// script. It also won't include each of stdout or stderr in the error string if they are used in
-/// the script.
+/// When called, the child process is spawned and a Child-typed value is returned, which supports:
+/// * Indices:
+///   * `:stdin`: A function which may be passed an `Into ByteStream` argument to write to the
+///   child's stdin, or a `Unit` argument to close the child's stdin.
+///   * `:stdout`: The standard output `ByteStream` of the child.
+///   * `:stderr`: The standard error `ByteStream` of the child.
+///   * `:exit`: The exit status of the child (waiting for the child to terminate), as an `ExitStatus`.
+///   * `:success`: A convenience index to wait for the child to complete, evaluating to `Unit` on
+///   successful exit or an `Error` indicating the child's exit status, stderr, stdout, and command
+///   line. __This is the nested value that is evaluated if a `Child` type is evaluated in sequence
+///   in a block.__
+/// * Conversion to `Bool` and `Number`, which is the same as converting `exit` to the type (exit
+/// success/failure and exit code, respectively).
+/// * Conversion to `String` and `ByteStream`, which is the same as converting `stdout` to the type
+/// _after_ evaluating the `success` index to ensure no error occurred.
 ///
 /// The `ExitStatus` type can be converted to `Bool` to check whether the program returned a
 /// successful status. It can also be converted to `Number`; if the process exited as the result of
@@ -129,7 +152,6 @@ pub async fn function(
     cmd: _,
     (env): [types::Map],
     (pwd): [_],
-    (stdin): [_],
     (retain_terminal): [_],
     ...
 ) -> Value {
@@ -164,11 +186,6 @@ pub async fn function(
         None => None,
     };
 
-    let stdin = match stdin {
-        Some(v) => Some(traits::into::<types::ByteStream>(v).await?),
-        None => None,
-    };
-
     let retain_terminal = match retain_terminal {
         Some(v) => traits::into::<types::Bool>(v).await?.as_ref().0,
         None => false,
@@ -199,9 +216,43 @@ pub async fn function(
         command.current_dir(v.as_ref().as_ref());
     }
 
-    if stdin.is_some() {
-        command.stdin(Stdio::piped());
-    }
+    let command_string = {
+        fn string_lit(s: &mut String, v: &std::ffi::OsStr) {
+            let v = v.to_string_lossy();
+            s.reserve(2 + v.len());
+            s.push('"');
+            s.extend(v.escape_default());
+            s.push('"');
+        }
+        let mut s = String::new();
+        s += "{pwd=";
+        match command.get_current_dir() {
+            None => s += "<none>",
+            Some(p) => s += &p.display().to_string(),
+        }
+        s += ",env={";
+        let mut first = true;
+        for (k, v) in command.get_envs() {
+            if !first {
+                s += ",";
+            }
+            string_lit(&mut s, k);
+            if let Some(v) = v {
+                s += "=";
+                string_lit(&mut s, v);
+            }
+            first = false;
+        }
+        s += "} ";
+        string_lit(&mut s, command.get_program());
+        for a in command.get_args() {
+            s += " ";
+            string_lit(&mut s, a);
+        }
+        s
+    };
+
+    command.stdin(Stdio::piped());
     command.stdout(Stdio::piped());
     command.stderr(Stdio::piped());
 
@@ -210,190 +261,89 @@ pub async fn function(
         disown_pgroup(&mut command);
     }
 
-    // Create channels for outputs
-    let (send_stdout, rcv_stdout) = channel();
-    let (send_stderr, rcv_stderr) = channel();
-    let (send_status, rcv_status) = channel();
+    log.debug(format_args!("spawning child process: {}", &command_string));
 
-    let args_source = ARGS_SOURCE.clone();
-    let run_command = async move {
-        log.debug(format!("spawning child process: {:?}", command));
+    let mut child = command
+        .spawn()
+        .add_primary_label(ARGS_SOURCE.with("while spawning this process"))?;
 
-        let mut child = command.spawn()
-            .add_primary_label(args_source.with("while spawning this process"))?;
-
-        // Handle stdin
-        let stdin = stdin.map(|v| v.as_ref().read());
-        let input = if let (Some(input), Some(mut v)) = (child.stdin.take(), stdin) {
-            let mut b = Blocking::new(input);
-            async move { ergo_runtime::io::copy_interactive(&mut v, &mut b).await }.boxed()
-        } else {
-            futures::future::ok(0).boxed()
-        };
-
-        // Handle stdout
-        let (mut out_pipe_send, out_pipe_recv) = pipe();
-        let mut cstdout = Blocking::new(child.stdout.take().unwrap());
-        let output = ergo_runtime::io::copy_interactive(&mut cstdout, &mut out_pipe_send);
-        let out_pipe_recv = if !send_stdout.is_canceled() {
-            drop(send_stdout.send(out_pipe_recv));
-            None
-        } else {
-            Some(out_pipe_recv)
-        };
-
-        // Handle stderr
-        let (mut err_pipe_send, err_pipe_recv) = pipe();
-        let mut cstderr = Blocking::new(child.stderr.take().unwrap());
-        let error = io::copy_interactive(&mut cstderr, &mut err_pipe_send);
-        let err_pipe_recv = if !send_stderr.is_canceled() {
-            drop(send_stderr.send(err_pipe_recv));
-            None
-        } else {
-            Some(err_pipe_recv)
-        };
-
-        // Handle running the child and getting the exit status
-        let exit_status =
-            Context::global()
-                .task
-                .spawn_blocking(move || {
-                    child.wait().map_err(|e| ergo_runtime::error! {
-                        labels: [ primary(ARGS_SOURCE.with("while waiting for process to exit")) ],
-                        error: e
-                    })
-                })
-                .map(|r| r.and_then(|v| v));
-
-        // Only run stdin while waiting for exit status
-        let exit_status = BoundTo::new(
-            exit_status,
-            input.map_err(|e| {
-                ergo_runtime::error! {
-                        labels: [ primary(ARGS_SOURCE.with("while copying stdin to this process")) ],
-                        error: e
-                }
-            }),
-        );
-
-        // Finally await everything
-        let exit_status = {
-            // TODO let _guard = super::task::ParentTask::remain_active();
-            futures::future::try_join3(
-                exit_status,
-                output.map_err(|e| {
-                    ergo_runtime::error! {
-                        labels: [ primary(ARGS_SOURCE.with("while copying stdout from this process")) ],
-                        error: e
-                    }
-                }),
-                error.map_err(|e| {
-                    ergo_runtime::error! {
-                        labels: [ primary(ARGS_SOURCE.with("while copying stderr from this process")) ],
-                        error: e
-                    }
-                }),
-            )
-            .await?
-            .0
-        };
-
-        log.debug(format!("child process exited: {:?}", command));
-
-        // These pipes should be dropped here so that, if an error occurred, the recv pipes
-        // terminate correctly if gathering stdout/stderr.
-        drop(out_pipe_send);
-        drop(err_pipe_send);
-
-        // Produce the final result
-        if send_status.is_canceled() {
-            if !exit_status.success() {
-                macro_rules! fetch_named {
-                    ( $output:expr, $name:expr ) => {
-                        if let Some(mut recv_pipe) = $output {
-                            let mut s = String::new();
-                            use futures::io::AsyncReadExt;
-                            if recv_pipe.read_to_string(&mut s).await.is_ok() {
-                                if !s.is_empty() {
-                                    Some(format!("{}: {}", $name, s))
-                                } else {
-                                    None
-                                }
-                            } else {
-                                None
-                            }
-                        } else {
-                            None
-                        }
-                        .unwrap_or_default()
-                    };
-                }
-                let stdout = fetch_named!(out_pipe_recv, "stdout");
-                let stderr = fetch_named!(err_pipe_recv, "stderr");
-                let mut diag = ergo_runtime::error::Diagnostic::from("command returned failure exit status")
-                    .add_primary_label(ARGS_SOURCE.with(""));
-                if !stdout.is_empty() {
-                    diag = diag.add_note(stdout);
-                }
-                if !stderr.is_empty() {
-                    diag = diag.add_note(stderr);
-                }
-                return Err(diag.into());
+    // Handle stdin
+    let stdin = std::sync::Arc::new(Mutex::new(Some(Blocking::new(child.stdin.take().unwrap()))));
+    let stdin = types::ergo_fn_value! {
+        #[depends(^CALL_DEPENDS.clone())]
+        #[cloning(stdin)]
+        /// Send data to stdin of the child process.
+        ///
+        /// Arguments: `:value`
+        ///
+        /// If `value` is `Into ByteStream`, converts the value and copies the ByteStream to the
+        /// child's stdin stream.
+        ///
+        /// If `value` is `Unit`, closes the stdin stream. Subsequent attempts to call this
+        /// function will produce an error.
+        async fn stdin(mut value: _) -> Value {
+            let mut stdin = stdin.lock().await;
+            if stdin.is_none() {
+                Err(ergo_runtime::error! {
+                    labels: [primary(ARGS_SOURCE.with(""))],
+                    error: "called `stdin` after closing"
+                })?;
             }
-        } else {
-            drop(send_status.send(exit_status));
+
+            Context::eval(&mut value).await?;
+            if value.is_type::<types::Unit>() {
+                stdin.take().unwrap().close().await?;
+            } else {
+                let stream = traits::into::<types::ByteStream>(value).await?;
+                io::copy_interactive(&mut stream.as_ref().read(), stdin.as_mut().unwrap()).await.into_diagnostic()?;
+            }
+            types::Unit.into()
         }
-
-        ergo_runtime::Result::Ok(())
-    }
-    .shared();
-
-    // Create output values
-    let complete = {
-        let run_command = run_command.clone();
-        Value::dynamic(
-            || async move {
-                try_result!(run_command.await);
-                types::Unit.into()
-            },
-            depends![dyn ^CALL_DEPENDS.clone(), nsid!(exec::complete)],
-        )
     };
 
+    // Handle stdout
     let stdout = Value::with_id(
-        types::ByteStream::new(ReadWhile::new(
-            ChannelRead::new(rcv_stdout),
-            run_command.clone(),
-        )),
+        types::ByteStream::new(Blocking::new(child.stdout.take().unwrap())),
         depends![dyn ^CALL_DEPENDS.clone(), nsid!(exec::stdout)],
     );
 
+    // Handle stderr
     let stderr = Value::with_id(
-        types::ByteStream::new(ReadWhile::new(
-            ChannelRead::new(rcv_stderr),
-            run_command.clone(),
-        )),
+        types::ByteStream::new(Blocking::new(child.stderr.take().unwrap())),
         depends![dyn ^CALL_DEPENDS.clone(), nsid!(exec::stderr)],
     );
 
-    let rcv_status = rcv_status.shared();
+    // Handle running the child and getting the exit status
+    let cs = command_string.clone();
+    let exit_status = Context::global()
+        .task
+        .spawn_blocking(move || {
+            let ret = child.wait().map_err(|e| {
+                ergo_runtime::error! {
+                    labels: [ primary(ARGS_SOURCE.with("while waiting for process to exit")) ],
+                    error: e
+                }
+            });
+            log.debug(format_args!("child process exited ({:?}): {}", ret, &cs));
+            ret
+        })
+        .map(|r| r.and_then(|v| v))
+        .shared();
     let exit_status = Value::dynamic(
-        move || async move {
-            try_result!(run_command.await);
-            ExitStatus::from(try_result!(rcv_status.await.add_primary_label(
-                args_source.with("while retrieving exit status of this process")
-            )))
-            .into()
-        },
-        depends![dyn ^CALL_DEPENDS, nsid!(exec::exit_status)],
+        move || async move { ExitStatus::from(try_result!(exit_status.await)).into() },
+        depends![dyn ^CALL_DEPENDS.clone(), nsid!(exec::exit_status)],
     );
 
-    crate::make_string_map! { source ARGS_SOURCE,
-        "stdout" = stdout,
-        "stderr" = stderr,
-        "exit-status" = exit_status,
-        "complete" = complete
-    }
+    Value::with_id(
+        Child {
+            command_string: command_string.into(),
+            stdin,
+            stdout,
+            stderr,
+            exit_status,
+        },
+        depends![dyn ^CALL_DEPENDS, nsid!(exec::child)],
+    )
 }
 
 #[cfg(unix)]
@@ -418,6 +368,85 @@ ergo_runtime::type_system::ergo_traits_fn! {
     traits::IntoTyped::<CommandString>::add_impl::<types::String>(traits);
     traits::IntoTyped::<CommandString>::add_impl::<types::Path>(traits);
     ergo_runtime::ergo_type_name!(traits, CommandString);
+
+    // Child traits
+    ergo_runtime::ergo_type_name!(traits, Child);
+    impl traits::IntoTyped<types::Bool> for Child {
+        async fn into_typed(self) -> Value {
+            traits::into::<types::Bool>(self.as_ref().exit_status.clone()).await.into()
+        }
+    }
+    impl traits::IntoTyped<types::Number> for Child {
+        async fn into_typed(self) -> Value {
+            traits::into::<types::Number>(self.as_ref().exit_status.clone()).await.into()
+        }
+    }
+    impl traits::IntoTyped<types::String> for Child {
+        async fn into_typed(self) -> Value {
+            try_result!(traits::bind_no_error(self.clone().into(), types::Index(types::String::from(CHILD_SUCCESS_INDEX).into()).into()).await);
+            traits::into::<types::String>(self.as_ref().stdout.clone()).await.into()
+        }
+    }
+    impl traits::IntoTyped<types::ByteStream> for Child {
+        async fn into_typed(self) -> Value {
+            try_result!(traits::bind_no_error(self.clone().into(), types::Index(types::String::from(CHILD_SUCCESS_INDEX).into()).into()).await);
+            traits::into::<types::ByteStream>(self.as_ref().stdout.clone()).await.into()
+        }
+    }
+    impl traits::Nested for Child {
+        async fn nested(&self) -> RVec<Value> {
+            rvec![traits::bind(SELF_VALUE.clone(), types::Index(types::String::from(CHILD_SUCCESS_INDEX).into()).into()).await]
+        }
+    }
+    impl traits::Bind for Child {
+        async fn bind(&self, arg: Value) -> Value {
+            let src = Source::get(&arg);
+            let ind = try_result!(Context::eval_as::<types::Index>(arg).await).to_owned().0;
+            let ind = try_result!(Context::eval_as::<types::String>(ind).await);
+            let s = ind.as_ref().as_str();
+            match s {
+                "stdin" => self.stdin.clone(),
+                "stdout" => self.stdout.clone(),
+                "stderr" => self.stderr.clone(),
+                "exit" => self.exit_status.clone(),
+                "success" => {
+                    let command_string = self.command_string.clone();
+                    let stdout = self.stdout.clone();
+                    let stderr = self.stderr.clone();
+                    let exit_status = self.exit_status.clone();
+                    Value::dynamic(move || async move {
+                            let exit_status = try_result!(Context::eval_as::<ExitStatus>(exit_status).await);
+                            if exit_status.as_ref().success() {
+                                types::Unit.into()
+                            } else {
+                                let stdout = try_result!(traits::into::<types::String>(stdout).await);
+                                let stderr = try_result!(traits::into::<types::String>(stderr).await);
+
+                                ergo_runtime::error! {
+                                    labels: [primary(src.with(""))],
+                                    notes: [
+                                        format_args!("command was: {}", command_string),
+                                        format_args!("exit status was: {}", exit_status.as_ref()),
+                                        format_args!("stdout was: {}", stdout.as_ref()),
+                                        format_args!("stderr was: {}", stderr.as_ref())
+                                    ],
+                                    error: "command returned failure exit status"
+                                }.into()
+                            }
+                        },
+                        depends![dyn SELF_VALUE, nsid!(exec::success)]
+                    )
+                }
+                _ => {
+                    ergo_runtime::error! {
+                        labels: [ primary(src.with("")) ],
+                        notes: [ "supported indices: `stdin`, `stdout`, `stderr`, `exit`, `success`" ],
+                        error: "unrecognized Child index"
+                    }.into()
+                }
+            }
+        }
+    }
 
     // ExitStatus traits
     traits::IntoTyped::<types::Bool>::add_impl::<ExitStatus>(traits);
@@ -444,239 +473,50 @@ ergo_runtime::type_system::ergo_traits_fn! {
     }
 }
 
-type PipeBuf = std::io::Cursor<Box<[u8]>>;
-
-struct PipeSend {
-    send: futures::channel::mpsc::UnboundedSender<tokio::io::Result<PipeBuf>>,
-}
-
-type PipeRecv = tokio_util::io::StreamReader<
-    futures::stream::Fuse<futures::channel::mpsc::UnboundedReceiver<tokio::io::Result<PipeBuf>>>,
-    PipeBuf,
->;
-
-fn pipe() -> (PipeSend, io::Tokio<PipeRecv>) {
-    let (send, recv) = futures::channel::mpsc::unbounded::<io::Result<PipeBuf>>();
-    (
-        PipeSend { send },
-        io::Tokio::new(tokio_util::io::StreamReader::new(recv.fuse())),
-    )
-}
-
-impl io::AsyncWrite for PipeSend {
-    fn poll_write(
-        self: std::pin::Pin<&mut Self>,
-        cx: &mut std::task::Context,
-        buf: &[u8],
-    ) -> std::task::Poll<io::Result<usize>> {
-        use futures::sink::Sink;
-        use std::task::Poll::*;
-        let mut me = std::pin::Pin::new(&mut self.get_mut().send);
-        match me.as_mut().poll_ready(cx) {
-            Pending => return Pending,
-            Ready(Err(_)) => return Ready(Err(std::io::ErrorKind::BrokenPipe.into())),
-            Ready(Ok(())) => (),
-        }
-
-        Ready(
-            if me
-                .start_send(Ok(std::io::Cursor::new(buf.to_vec().into_boxed_slice())))
-                .is_err()
-            {
-                Err(std::io::ErrorKind::BrokenPipe.into())
-            } else {
-                Ok(buf.len())
-            },
-        )
-    }
-
-    fn poll_flush(
-        self: std::pin::Pin<&mut Self>,
-        cx: &mut std::task::Context,
-    ) -> std::task::Poll<io::Result<()>> {
-        use futures::sink::Sink;
-        let me = std::pin::Pin::new(&mut self.get_mut().send);
-        me.poll_flush(cx)
-            .map(|r| r.map_err(|_| std::io::ErrorKind::BrokenPipe.into()))
-    }
-
-    fn poll_close(
-        self: std::pin::Pin<&mut Self>,
-        cx: &mut std::task::Context,
-    ) -> std::task::Poll<io::Result<()>> {
-        use futures::sink::Sink;
-        let me = std::pin::Pin::new(&mut self.get_mut().send);
-        me.poll_close(cx)
-            .map(|r| r.map_err(|_| std::io::ErrorKind::BrokenPipe.into()))
-    }
-}
-
-enum ChannelRead<R> {
-    Waiting(futures::channel::oneshot::Receiver<R>),
-    Read(R),
-}
-
-impl<R: io::AsyncRead + std::marker::Unpin> ChannelRead<R> {
-    pub fn new(channel: futures::channel::oneshot::Receiver<R>) -> Self {
-        ChannelRead::Waiting(channel)
-    }
-}
-
-impl<R> io::AsyncRead for ChannelRead<R>
-where
-    R: io::AsyncRead + std::marker::Unpin,
-{
-    fn poll_read(
-        mut self: std::pin::Pin<&mut Self>,
-        cx: &mut std::task::Context,
-        buf: &mut [u8],
-    ) -> std::task::Poll<io::Result<usize>> {
-        let me = &mut *self;
-        use futures::future::Future;
-        use std::task::Poll::*;
-        loop {
-            let new_me = match me {
-                ChannelRead::Waiting(r) => match std::pin::Pin::new(r).poll(cx) {
-                    Pending => return Pending,
-                    Ready(Err(_)) => return Ready(Ok(0)),
-                    Ready(Ok(v)) => ChannelRead::Read(v),
-                },
-                ChannelRead::Read(r) => return std::pin::Pin::new(r).poll_read(cx, buf),
-            };
-            *me = new_me;
-        }
-    }
-}
-
-struct ReadWhile<R, Fut> {
-    read: R,
-    read_done: bool,
-    fut: Fut,
-    fut_done: bool,
-}
-
-impl<R, Fut> ReadWhile<R, Fut> {
-    pub fn new(read: R, fut: Fut) -> Self {
-        ReadWhile {
-            read,
-            read_done: false,
-            fut,
-            fut_done: false,
-        }
-    }
-}
-
-impl<T, R, Fut> io::AsyncRead for ReadWhile<R, Fut>
-where
-    R: io::AsyncRead + std::marker::Unpin,
-    Fut: futures::future::Future<Output = Result<T, Error>> + std::marker::Unpin,
-{
-    fn poll_read(
-        mut self: std::pin::Pin<&mut Self>,
-        cx: &mut std::task::Context,
-        buf: &mut [u8],
-    ) -> std::task::Poll<io::Result<usize>> {
-        use std::task::Poll::*;
-        let result = if !self.read_done {
-            let me = &mut *self;
-            match std::pin::Pin::new(&mut me.read).poll_read(cx, buf) {
-                Ready(Err(e)) => return Ready(Err(e)),
-                Ready(Ok(0)) => {
-                    me.read_done = true;
-                    Ready(Ok(0))
-                }
-                otherwise => otherwise,
-            }
-        } else {
-            Pending
-        };
-        if !self.fut_done {
-            let me = &mut *self;
-            match std::pin::Pin::new(&mut me.fut).poll(cx) {
-                Ready(Err(e)) => return Ready(Err(e.into())),
-                Ready(Ok(_)) => {
-                    me.fut_done = true;
-                }
-                _ => (),
-            }
-        };
-
-        if self.fut_done && self.read_done {
-            Ready(Ok(0))
-        } else if let Ready(Ok(0)) = result {
-            Pending
-        } else {
-            result
-        }
-    }
-}
-
-struct BoundTo<MasterFut, Fut> {
-    master: MasterFut,
-    other: Option<Fut>,
-}
-
-impl<M, Fut> BoundTo<M, Fut> {
-    pub fn new(master: M, other: Fut) -> Self {
-        BoundTo {
-            master,
-            other: Some(other),
-        }
-    }
-}
-
-impl<MasterFut, Fut, MT, FT, E> futures::future::Future for BoundTo<MasterFut, Fut>
-where
-    MasterFut: futures::future::Future<Output = Result<MT, E>> + std::marker::Unpin,
-    Fut: futures::future::Future<Output = Result<FT, E>> + std::marker::Unpin,
-{
-    type Output = MasterFut::Output;
-
-    fn poll(
-        mut self: std::pin::Pin<&mut Self>,
-        cx: &mut std::task::Context,
-    ) -> std::task::Poll<Self::Output> {
-        let me = &mut *self;
-        use std::task::Poll::*;
-        if let Some(o) = &mut me.other {
-            match std::pin::Pin::new(o).poll(cx) {
-                Ready(Err(e)) => return Ready(Err(e)),
-                Ready(Ok(_)) => {
-                    me.other = None;
-                }
-                Pending => (),
-            }
-        }
-
-        std::pin::Pin::new(&mut me.master).poll(cx)
-    }
-}
-
 // Only run these tests on unix, where we assume the existence of some programs.
 #[cfg(all(test, unix))]
 mod test {
     ergo_script::tests! {
         fn exec(t) {
-            t.assert_success("self:exec echo hello |>:complete")
+            t.assert_success("self:exec echo hello; ()");
+            t.assert_success("self:exec echo hello |>:success");
         }
 
         fn exec_fail(t) {
-            t.assert_fail("self:exec false |>:complete")
+            t.assert_fail("self:exec false; ()");
+            t.assert_fail("self:exec false |>:success");
         }
 
         fn exit_status_number(t) {
-            t.assert_eq("self:Number:from <| self:exec false |>:exit-status", "self:Number:new 1");
-            t.assert_eq("self:Number:from <| self:exec true |>:exit-status", "self:Number:new 0");
+            t.assert_eq("self:Number:from <| self:exec false |>:exit", "self:Number:new 1");
+            t.assert_eq("self:Number:from <| self:exec true |>:exit", "self:Number:new 0");
+            t.assert_eq("self:Number:from <| self:exec false", "self:Number:new 1");
+            t.assert_eq("self:Number:from <| self:exec true", "self:Number:new 0");
+        }
+
+        fn exit_bool(t) {
+            t.assert_eq("self:Bool:from <| self:exec false |>:exit", "self:Bool:false");
+            t.assert_eq("self:Bool:from <| self:exec true |>:exit", "self:Bool:true");
+            t.assert_eq("self:Bool:from <| self:exec false", "self:Bool:false");
+            t.assert_eq("self:Bool:from <| self:exec true", "self:Bool:true");
         }
 
         fn exec_unset_args(t) {
             t.assert_eq("self:String:from <| self:exec echo $unset hello $unset world |>:stdout", "\"hello world\\n\"");
-            t.assert_fail("self:exec $unset echo |>:complete");
+            t.assert_fail("self:exec $unset echo |>:success");
         }
 
         fn exec_unit_args(t) {
             t.assert_eq("self:String:from <| self:exec echo () hello () world |>:stdout", "\"hello world\\n\"");
-            t.assert_fail("self:exec () echo |>:complete");
+            t.assert_fail("self:exec () echo |>:success");
+        }
+
+        fn stdin(t) {
+            t.assert_eq("child = self:exec cat; child:stdin hello; child:stdin (); self:String:from child:stdout", "\"hello\"");
+        }
+
+        fn to_string(t) {
+            t.assert_eq("self:String:from <| self:exec echo abc", "\"abc\\n\"");
         }
     }
 }